requirements.txt
fastapi
celery
redis
uvicorn
python-multipart
aiofiles
filetype
pillow